Use index files to organize ts imports
