+++ src/__tests__/hooks/use-ynab-token/use-ynab-token.test.tsx
@@ -1,21 +1,19 @@
 import { renderHook } from '@testing-library/react';
 import { useRouter } from 'next/navigation';
-import getYnabTokenFromUrlHash from '@/hooks/use-ynab-token/helpers/get-ynab-token-from-url-hash';
-import setYnabTokenAsCookie from '@/hooks/use-ynab-token/helpers/set-ynab-token-as-cookie';
-import useYnabToken from '@/hooks/use-ynab-token/use-ynab-token';
+import {
+  getYnabTokenFromUrlHash,
+  setYnabTokenAsCookie,
+} from '@/hooks/use-ynab-token/helpers';
+import { useYnabToken } from '@/hooks/use-ynab-token';
 
 jest.mock('next/navigation', () => ({
   useRouter: jest.fn(),
 }));
 
-jest.mock(
-  '../../../hooks/use-ynab-token/helpers/get-ynab-token-from-url-hash',
-  () => jest.fn()
-);
-jest.mock(
-  '../../../hooks/use-ynab-token/helpers/set-ynab-token-as-cookie',
-  () => jest.fn()
-);
+jest.mock('hooks/use-ynab-token/helpers', () => ({
+  getYnabTokenFromUrlHash: jest.fn(),
+  setYnabTokenAsCookie: jest.fn(),
+}));
 
 describe('useYnabToken', () => {
   const mockPush = jest.fn();
