+++ src/hooks/use-ynab-token/helpers/set-ynab-token-as-cookie.ts
@@ -5,11 +5,12 @@ import cookieCutter from '@boiseitguru/cookie-cutter';
  * @param accessToken The YNAB access token to set as a cookie.
  * @param expiresIn The expiration time of the access token in seconds.
  */
-const setYnabTokenAsCookie = (accessToken: string, expiresIn: number) => {
+export const setYnabTokenAsCookie = (
+  accessToken: string,
+  expiresIn: number
+) => {
   const cookieExpiration = new Date(
     new Date().getTime() + expiresIn * 1000
   ).toUTCString();
   cookieCutter.set('ynabToken', accessToken, { expires: cookieExpiration });
 };
-
-export default setYnabTokenAsCookie;
