+++ src/hooks/use-ynab-token/index.ts
@@ -1,0 +1,1 @@
+export { useYnabToken } from './use-ynab-token';
